api_server/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
apt-clone==0.2.1
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
asyncio==3.4.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.6.0
bidict==0.22.1
black==23.3.0
bleach==6.0.0
blinker==1.4
Brlapi==0.6.6
certifi==2018.1.18
cffi==1.15.1
chardet==3.0.4
click==8.1.7
comm==0.1.4
cryptography==2.1.4
cupshelpers==1.0
debugpy==1.7.0
decorator==4.1.2
defer==1.0.6
defusedxml==0.7.1
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.4
exceptiongroup==1.2.0
fastapi==0.103.2
fastjsonschema==2.19.1
feedparser==5.2.1
h11==0.14.0
html5lib==0.999999999
httplib2==0.9.2
httptools==0.6.0
idna==3.6
importlib==1.0.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
jedi==0.19.1
Jetson.GPIO==2.0.11
Jinja2==3.1.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.5.8
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.13.3
oauth==1.0.1
oauthlib==2.0.6
onboard==1.4.1
packaging==23.2
PAM==0.4.2
pandas==0.22.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pbr==6.0.0
pexpect==4.9.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.43
protobuf==3.0.0
psutil==5.9.7
ptyprocess==0.7.0
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyGObject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
pyrsistent==0.19.3
pyserial==3.5
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.2
python-debian==0.1.32
python-dotenv==0.21.1
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2018.3
pyxattr==0.6.0
pyxdg==0.25
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.1
requests==2.18.4
requests-unixsocket==0.1.5
scipy==0.19.1
SecretStorage==2.3.1
Send2Trash==1.8.2
simple-websocket==1.0.0
simplejson==3.13.2
six==1.11.0
sniffio==1.3.0
ssh-import-id==5.7
starlette==0.27.0
system-service==0.3
systemd-python==234
terminado==0.17.1
testresources==2.0.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
typed-ast==1.5.5
typing_extensions==4.7.1
ubuntu-drivers-common==0.0.0
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
urwid==2.0.1
uvicorn==0.22.0
uvloop==0.18.0
wadllib==1.3.2
watchfiles==0.20.0
wcwidth==0.2.12
webencodings==0.5
websocket-client==1.6.1
websockets==11.0.3
widgetsnbextension==4.0.9
wsproto==1.2.0
xkit==0.0.0
youtube_dl==2018.3.14
zipp==3.15.0
zope.interface==4.3.2
